refactor: small consolidation of functions
