+++ lib/dispatcher/client-h1.js
@@ -47,7 +47,6 @@ const {
   kMaxRequests,
   kCounter,
   kMaxResponseSize,
-  kListeners,
   kOnError,
   kResume,
   kHTTPContext
@@ -56,22 +55,10 @@ const {
 const constants = require('../llhttp/constants.js')
 const EMPTY_BUF = Buffer.alloc(0)
 const FastBuffer = Buffer[Symbol.species]
+const addListener = util.addListener
+const removeAllListeners = util.removeAllListeners
 
 let extractBody
-
-function addListener (obj, name, listener) {
-  const listeners = (obj[kListeners] ??= [])
-  listeners.push([name, listener])
-  obj.on(name, listener)
-  return obj
-}
-
-function removeAllListeners (obj) {
-  for (const [name, listener] of obj[kListeners] ?? []) {
-    obj.removeListener(name, listener)
-  }
-  obj[kListeners] = null
-}
 
 async function lazyllhttp () {
   const llhttpWasmData = process.env.JEST_WORKER_ID ? require('../llhttp/llhttp-wasm.js') : undefined
@@ -719,14 +706,14 @@ async function connectH1 (client, socket) {
       const requests = client[kQueue].splice(client[kRunningIdx])
       for (let i = 0; i < requests.length; i++) {
         const request = requests[i]
-        errorRequest(client, request, err)
+        util.errorRequest(client, request, err)
       }
     } else if (client[kRunning] > 0 && err.code !== 'UND_ERR_INFO') {
       // Fail head of pipeline.
       const request = client[kQueue][client[kRunningIdx]]
       client[kQueue][client[kRunningIdx]++] = null
 
-      errorRequest(client, request, err)
+      util.errorRequest(client, request, err)
     }
 
     client[kPendingIdx] = client[kRunningIdx]
@@ -831,15 +818,6 @@ function resumeH1 (client) {
   }
 }
 
-function errorRequest (client, request, err) {
-  try {
-    request.onError(err)
-    assert(request.aborted)
-  } catch (err) {
-    client.emit('error', err)
-  }
-}
-
 // https://www.rfc-editor.org/rfc/rfc7230#section-3.3.2
 function shouldSendContentLength (method) {
   return method !== 'GET' && method !== 'HEAD' && method !== 'OPTIONS' && method !== 'TRACE' && method !== 'CONNECT'
@@ -906,7 +884,7 @@ function writeH1 (client, request) {
   // A user agent may send a Content-Length header with 0 value, this should be allowed.
   if (shouldSendContentLength(method) && contentLength > 0 && request.contentLength !== null && request.contentLength !== contentLength) {
     if (client[kStrictContentLength]) {
-      errorRequest(client, request, new RequestContentLengthMismatchError())
+      util.errorRequest(client, request, new RequestContentLengthMismatchError())
       return false
     }
 
@@ -921,12 +899,12 @@ function writeH1 (client, request) {
         return
       }
 
-      errorRequest(client, request, err || new RequestAbortedError())
+      util.errorRequest(client, request, err || new RequestAbortedError())
 
       util.destroy(socket, new InformationalError('aborted'))
     })
   } catch (err) {
-    errorRequest(client, request, err)
+    util.errorRequest(client, request, err)
   }
 
   if (request.aborted) {
