+++ lib/dispatcher/client-h2.js
@@ -22,7 +22,6 @@ const {
   kSocket,
   kStrictContentLength,
   kOnError,
-  // HTTP2
   kMaxConcurrentStreams,
   kHTTP2Session,
   kResume
@@ -92,11 +91,12 @@ async function connectH2 (client, socket) {
   session[kOpenStreams] = 0
   session[kClient] = client
   session[kSocket] = socket
-  session.on('error', onHttp2SessionError)
-  session.on('frameError', onHttp2FrameError)
-  session.on('end', onHttp2SessionEnd)
-  session.on('goaway', onHTTP2GoAway)
-  session.on('close', function () {
+
+  util.addListener(session, 'error', onHttp2SessionError)
+  util.addListener(session, 'frameError', onHttp2FrameError)
+  util.addListener(session, 'end', onHttp2SessionEnd)
+  util.addListener(session, 'goaway', onHTTP2GoAway)
+  util.addListener(session, 'close', function () {
     const { [kClient]: client } = this
 
     const err = this[kError] || new SocketError('closed', util.getSocketInfo(this))
@@ -109,7 +109,7 @@ async function connectH2 (client, socket) {
     const requests = client[kQueue].splice(client[kRunningIdx])
     for (let i = 0; i < requests.length; i++) {
       const request = requests[i]
-      errorRequest(client, request, err)
+      util.errorRequest(client, request, err)
     }
 
     client[kPendingIdx] = client[kRunningIdx]
@@ -120,19 +120,21 @@ async function connectH2 (client, socket) {
 
     client[kResume]()
   })
+
   session.unref()
 
   client[kHTTP2Session] = session
   socket[kHTTP2Session] = session
 
-  socket.on('error', function (err) {
+  util.addListener(socket, 'error', function (err) {
     assert(err.code !== 'ERR_TLS_CERT_ALTNAME_INVALID')
 
     this[kError] = err
 
     this[kClient][kOnError](err)
   })
-  socket.on('end', function () {
+
+  util.addListener(socket, 'end', function () {
     util.destroy(this, new SocketError('other side closed', util.getSocketInfo(this)))
   })
 
@@ -245,7 +247,7 @@ function writeH2 (client, request) {
   const { body, method, path, host, upgrade, expectContinue, signal, headers: reqHeaders } = request
 
   if (upgrade) {
-    errorRequest(client, request, new Error('Upgrade not supported for H2'))
+    util.errorRequest(client, request, new Error('Upgrade not supported for H2'))
     return false
   }
 
@@ -298,10 +300,10 @@ function writeH2 (client, request) {
         }
       }
 
-      errorRequest(client, request, err)
+      util.errorRequest(client, request, err)
     })
   } catch (err) {
-    errorRequest(client, request, err)
+    util.errorRequest(client, request, err)
   }
 
   if (method === 'CONNECT') {
@@ -376,7 +378,7 @@ function writeH2 (client, request) {
   // A user agent may send a Content-Length header with 0 value, this should be allowed.
   if (shouldSendContentLength(method) && contentLength > 0 && request.contentLength != null && request.contentLength !== contentLength) {
     if (client[kStrictContentLength]) {
-      errorRequest(client, request, new RequestContentLengthMismatchError())
+      util.errorRequest(client, request, new RequestContentLengthMismatchError())
       return false
     }
 
@@ -418,7 +420,7 @@ function writeH2 (client, request) {
     // as there's no value to keep it open.
     if (request.aborted || request.completed) {
       const err = new RequestAbortedError()
-      errorRequest(client, request, err)
+      util.errorRequest(client, request, err)
       util.destroy(stream, err)
       return
     }
@@ -452,13 +454,12 @@ function writeH2 (client, request) {
     }
 
     const err = new InformationalError('HTTP/2: stream half-closed (remote)')
-    errorRequest(client, request, err)
+    util.errorRequest(client, request, err)
     util.destroy(stream, err)
   })
 
   stream.once('close', () => {
     session[kOpenStreams] -= 1
-    // TODO(HTTP/2): unref only if current streams count is 0
     if (session[kOpenStreams] === 0) {
       session.unref()
     }
@@ -467,13 +468,14 @@ function writeH2 (client, request) {
   stream.once('error', function (err) {
     if (client[kHTTP2Session] && !client[kHTTP2Session].destroyed && !this.closed && !this.destroyed) {
       session[kOpenStreams] -= 1
+      util.errorRequest(client, request, err)
       util.destroy(stream, err)
     }
   })
 
   stream.once('frameError', (type, code) => {
     const err = new InformationalError(`HTTP/2: "frameError" received - type ${type}, code ${code}`)
-    errorRequest(client, request, err)
+    util.errorRequest(client, request, err)
 
     if (client[kHTTP2Session] && !client[kHTTP2Session].destroyed && !this.closed && !this.destroyed) {
       session[kOpenStreams] -= 1
