+++ lib/dispatcher/client.js
@@ -338,7 +338,7 @@ class Client extends DispatcherBase {
       const requests = this[kQueue].splice(this[kPendingIdx])
       for (let i = 0; i < requests.length; i++) {
         const request = requests[i]
-        errorRequest(this, request, err)
+        util.errorRequest(this, request, err)
       }
 
       const callback = () => {
@@ -378,7 +378,7 @@ function onError (client, err) {
     const requests = client[kQueue].splice(client[kRunningIdx])
     for (let i = 0; i < requests.length; i++) {
       const request = requests[i]
-      errorRequest(client, request, err)
+      util.errorRequest(client, request, err)
     }
     assert(client[kSize] === 0)
   }
@@ -502,7 +502,7 @@ async function connect (client) {
       assert(client[kRunning] === 0)
       while (client[kPending] > 0 && client[kQueue][client[kPendingIdx]].servername === client[kServerName]) {
         const request = client[kQueue][client[kPendingIdx]++]
-        errorRequest(client, request, err)
+        util.errorRequest(client, request, err)
       }
     } else {
       onError(client, err)
@@ -606,15 +606,6 @@ function _resume (client, sync) {
     } else {
       client[kQueue].splice(client[kPendingIdx], 1)
     }
-  }
-}
-
-function errorRequest (client, request, err) {
-  try {
-    request.onError(err)
-    assert(request.aborted)
-  } catch (err) {
-    client.emit('error', err)
   }
 }
